refactor: move CLI commands to commands package (#58)
