+++ internal/runtime/plugin/manifest.go
